Удалить функционал разрешений с удаленного сервера
Переконфигурировать все связанные с данным функционал возможности на функционал локальных разрешений

+++ src/kz/tms/model/authentication/AuthenticationPrincipal.kt
@@ -4,5 +4,5 @@ import io.ktor.auth.*
 
 class AuthenticationPrincipal(
     val username: String,
-    val power: Long
+    val power: Int
 ) : Principal
